Update version and refactor components for improved functionality
- Bump version from 1.0.0 to 1.0.1 in package.json.
- Refactor AppShell component for better readability and consistency in import statements.
- Enhance BreakoutGroupsCard by integrating a custom hook for clipboard functionality.
- Simplify QuizTimerCard by utilizing a new useTimer hook for timer management.
- Introduce QuizEditorForm and QuizImportCard components for improved quiz management and import functionality.
- Implement useCopyToClipboard hook for streamlined clipboard operations across components.
