+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "teacherbuddy",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "private": true,
   "scripts": {
     "dev": "bun --bun next dev",
